requirements.txt
asgiref==3.11.0
Django==5.2.9
sqlparse==0.5.4
typing_extensions==4.15.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1
drf-spectacular==0.29.0
